переместил функции из .cpp в .h файл и удвлил .cpp файл за ненадобностью
